01-cedar/photoapp-cedar/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
Pillow==10.4.0 
python-dotenv==1.0.1
cedarpy==4.1.0
pytest==8.0.2 